requirements.txt
mysql-connector-python
PySide6
pyqt
pyqt6-tools
python-dotenv
bycript
python-docx
openpyxl
Pillow
fpdf
